+++ client/web/src/site-admin/SiteAdminPingsPage.tsx
@@ -200,6 +200,20 @@ export const SiteAdminPingsPage: React.FunctionComponent<Props> = props => {
                         <li>Total number of tours closed</li>
                     </ul>
                 </li>
+                <li>
+                    Sourcegraph extension activation statistics
+                    <ul>
+                        <li>Total number of users that use a given non-default Sourcegraph extension</li>
+                        <li>
+                            Average number of activations for users that use a given non-default Sourcegraph extension
+                        </li>
+                        <li>Total number of users that use non-default Sourcegraph extensions</li>
+                        <li>
+                            Average number of non-default extensions enabled for users that use non-default Sourcegraph
+                            extensions
+                        </li>
+                    </ul>
+                </li>
             </ul>
             {updatesDisabled ? (
                 <p>All telemetry is disabled.</p>