+++ cmd/frontend/internal/app/updatecheck/handler.go
@@ -187,6 +187,7 @@ type pingRequest struct {
 	CodeIntelUsage       json.RawMessage `json:"codeIntelUsage"`
 	NewCodeIntelUsage    json.RawMessage `json:"newCodeIntelUsage"`
 	SearchUsage          json.RawMessage `json:"searchUsage"`
+	ExtensionsUsage      json.RawMessage `json:"extensionsUsage"`
 	InitialAdminEmail    string          `json:"initAdmin"`
 	TotalUsers           int32           `json:"totalUsers"`
 	HasRepos             bool            `json:"repos"`
@@ -288,6 +289,7 @@ type pingPayload struct {
 	Repositories         json.RawMessage `json:"repositories"`
 	SearchOnboarding     json.RawMessage `json:"search_onboarding"`
 	DependencyVersions   json.RawMessage `json:"dependency_versions"`
+	ExtensionsUsage      json.RawMessage `json:"extensions_usage"`
 	InstallerEmail       string          `json:"installer_email"`
 	AuthProviders        string          `json:"auth_providers"`
 	ExtServices          string          `json:"ext_services"`
@@ -370,6 +372,7 @@ func marshalPing(pr *pingRequest, hasUpdate bool, clientAddr string, now time.Ti
 		SearchOnboarding:     pr.SearchOnboarding,
 		InstallerEmail:       pr.InitialAdminEmail,
 		DependencyVersions:   pr.DependencyVersions,
+		ExtensionsUsage:      pr.ExtensionsUsage,
 		AuthProviders:        strings.Join(pr.AuthProviders, ","),
 		ExtServices:          strings.Join(pr.ExternalServices, ","),
 		BuiltinSignupAllowed: strconv.FormatBool(pr.BuiltinSignupAllowed),